モーダルウィンドウ周りをリファクタ
